requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
coveralls==1.3.0
dj-database-url==0.5.0
Django==2.0.3
django-cors-headers==2.3.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docopt==0.6.2
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
pipetools==0.3.2
psycopg2==2.7.4
Pygments==2.2.0
PyJWT==1.6.4
python-dotenv==0.8.2
pytz==2018.3
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
